requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0


# Bleeding edge Django
django==1.11.2

# Configuration
django-environ==0.4.3

# Images
Pillow==4.1.1

# Password storage
argon2-cffi==16.3.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.32.0

# Python-PostgreSQL Database Adapter
psycopg2==2.7.1

# Time zones support
pytz==2017.2

# Redis support
django-redis==4.8.0
redis>=2.10.5

# 
# Django Q cluster
django-q==0.8.0
# 
# 
# Django rest framework
djangorestframework==3.6.3
# 
# 
django-rest-auth==0.9.1
# 
# 
# Graph QL
graphene>=1.0
graphene-django>=1.3
# 

# Your custom requirements go here
